move firebase config and initilization to index.js
