+++ package.json
@@ -9,6 +9,7 @@
     "@fullcalendar/list": "^5.7.0",
     "@fullcalendar/react": "^5.7.0",
     "@fullcalendar/timegrid": "^5.7.0",
+    "@reduxjs/toolkit": "^1.5.1",
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
