+++ public/index.html
@@ -47,6 +47,7 @@
   <!-- <script src="https://unpkg.com/leaflet@1.7.1/dist/leaflet.js"
     integrity="sha512-XQoYMqMTK8LvdxXYG3nZ448hOEQiglfqkJs1NOQV44cWnUrBc8PkAOcXy20w0vlaXaVUearIOBhiXZ5V3ynxwA=="
     crossorigin=""></script> -->
+  <script src="../src/index.js"></script>
 </body>
 
 </html>
