+++ src/Utils/FirebaseGoogle.js
@@ -1,0 +1,35 @@
+import firebase from "firebase/app";
+import "firebase/auth";
+import { useEffect, useState } from "react";
+
+// function FirebaseGoogle() {
+//   var provider = new firebase.auth.GoogleAuthProvider();
+//   //   provider.setCustomParameters({
+//   //     login_hint: "user@example.com",
+//   //   });
+//   firebase
+//     .auth()
+//     .signInWithPopup(provider)
+//     .then((result) => {
+//       /** @type {firebase.auth.OAuthCredential} */
+//       var credential = result.credential;
+
+//       // This gives you a Google Access Token. You can use it to access the Google API.
+//       var token = credential.accessToken;
+//       // The signed-in user info.
+//       var user = result.user;
+//       // ...
+//     })
+//     .catch((error) => {
+//       // Handle Errors here.
+//       var errorCode = error.code;
+//       var errorMessage = error.message;
+//       // The email of the user's account used.
+//       var email = error.email;
+//       // The firebase.auth.AuthCredential type that was used.
+//       var credential = error.credential;
+//       // ...
+//     });
+// }
+
+// export default FirebaseGoogle;
